VSCODE-569 made a couple minor changes to fix bug
Some pipelines didn't have a "selector" property in their target property. I don't know how this is possible, and it happens very infrequently, but some of our pipelines for VSCODE were breaking as a result. This is now fixed, and I am able to successfully load all of our pipelines.