VSCODE-582: Mention users in bitbucket pull requests and comments
This is an alpha version for mentioning users. Mentioning someone adds their uuid in the required format in the comment input, so the user will not see the actual user name in the input which can be confusing. It is better than nothing though and we can improve it later.