VSCODE-622 Issues now refresh when the project is updated.
When the configuration associated with the selected Jira site is changed, the Jira context is refreshed, which refreshes the Jira Issues explorer. Unfortunately, there appears to be a race condition, where the refresh occurs before the site change has actually gone through. As a result, a timeout is used to delay the refresh; since we're not actually checking if the update has completed, technically this isn't guaranteed to work. Furthermore, it's not a clean way to deal with this type of issue...