VSCODE-706 JQL presets can now be selected
There is now a button which opens a modal with JQL preset options. This allows users who don't know what JQL is or how to write JQL queries the ability to view issues relevant to them in the sidebar.

There is also a link to an explanation of JQL on the JQL config.