VSCODE-1019 Created a hint notification system that suggests key actions in various contexts
Currently there are 6 actions that are being tracked:
1. Viewing the Jira issue screen
2. Pressing start work on the Jira issue screen
3. Viewing the create issue screen
4. Pressing edit this file from a diff view
5. Viewing a pull request webview
6. Viewing a file diff

This is achieved using a HintProvider class, which manages most of the stuff you don't want to think about surrounding global state. The showHintNotification() method will look through all the actions saved to this HintProvider, filter out the ones that have already been done, and choose a random one. Since one piece of code is responsible for all hint notificaitons, we could easily add additional parameters like a probability to show a hint, or add far more hints in the future.

I also added a checklistExplorer, which looks like the other explores (e.g. Jira Issue Explorer). It periodically (once an hour, but this can be adjusted) displays a relavant hint). Right now this explorer is empty, but the idea is to fill this explorer with nodes that would contain key actions, and clicking them would open a tutorial page that demonstrates how to do that action with gifs (originally, each action was to have its own page, but upon reflection is makes a lot more sense to have one page and put each explanation in a panel/section much like what we do with the settings buttons that all open different sections of the config page).

Note: originally, the completed checklist was stored in the config, but this was leading to problems (and also, users shouldn't be able to see these as options), so now all of this is handled in the global store via Container.updateChecklistItem() and Container.getChecklistItem().